voiceops/voiceops-backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic>=2.0.0
prometheus-client>=0.20.0
python-multipart>=0.0.6
openai>=1.12.0
python-dotenv>=1.0.0